% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbar.R
\name{forecast.xgbar}
\alias{forecast.xgbar}
\title{Forecasting using xgboost models}
\usage{
\method{forecast}{xgbar}(object, h = NULL, xreg = NULL, ...)
}
\arguments{
\item{object}{An object of class "\code{xgbar}".  Usuall the result of a call to \code{\link{xgbar}}.}

\item{h}{Number of periods for forecasting}

\item{xreg}{Future values of regression variables.}

\item{...}{Ignored.}
}
\value{
An object of class \code{forecast}
}
\description{
Returns forecasts and other information for xgboost timeseries modesl fit with \code{xbgts}
}
\examples{
# Australian monthly gas production
gas_model <- xgbar(gas)
summary(gas_model)
gas_fc <- forecast(gas_model, h = 12)
plot(gas_fc)
}
\author{
Peter Ellis
}
\seealso{
\code{\link{xgbar}}, \code{\link[forecast]{forecast}}
}

