% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastxgb.R
\docType{package}
\name{forecastxgb-package}
\alias{forecastxgb}
\alias{forecastxgb-package}
\title{Extreme gradient boosting time series forecasting}
\description{
The \code{forecastxgb} package provides time series modelling and forecasting functions that combine 
the machine learning approach of Chen, He and Benesty's \code{xgboost}
with the convenient handling of time series and familiar API of Rob Hyndman's \code{forecast}
It applies to time series the Extreme Gradient Boosting proposed in \emph{Greedy Function Approximation: A 
Gradient Boosting Machine*, by Jermoe Friedman in 2001} (http://www.jstor.org/stable/2699986).
}

