% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{importance}
\alias{importance}
\title{Show importance of features in a xgbts model}
\usage{
importance(object, ...)
}
\arguments{
\item{object}{An object of class \code{xgbts}, usually created with \code{xgbts()}}

\item{...}{Extra parameters passed through to \code{xgb.importance}}
}
\value{
A \code{data.table} of the features used in the model with their average gain 
(and their weight for boosted tree model) in the model.
}
\description{
This is a light wrapper for \code{xgboost::xbg.importance} to make it easier to use with objects of class \code{xgbts}
}
\seealso{
\code{\link[xgboost]{xgb.importance}}
}

