% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\docType{data}
\name{Mcomp_results}
\alias{Mcomp_results}
\title{M3 forecasting results}
\format{A data frame with 75 rows and three columns.}
\usage{
Mcomp_results
}
\description{
Summary data from four models, and 11 combinations of models, against the data from the M3 forecasting competition.
}
\details{
Full details are in the vignette of how a similar series with tourism competition data was generated.  
The data shows the average mean absolute scaled error 
(MASE) from using \code{xgbar} (x), \code{auto.arima} (a), \code{nnetar} (n) and \code{thetaf} (f) to 
generate forecasts of 3,003 data series from a range of sectors, in the M3 forecasting competition.

\itemize{
\item MASE A mean mean absolute squared error
\item model model, or ensemble of models, to which the MASE applies
\item Frequency The frequency of the subset of data from which the mean MASE was calculated.
}
}
\examples{
if(require(ggplot2)){
leg <- "f: Theta; forecast::thetaf\\na: ARIMA; forecast::auto.arima
n: Neural network; forecast::nnetar\\nx: Extreme gradient boosting; forecastxgb::xgbar"

ggplot(Mcomp_results, aes(x = model, y =  MASE, colour = Frequency, label = model)) +
  geom_text(size = 4) +
  geom_line(aes(x = as.numeric(model)), alpha = 0.25) +
  annotate("text", x = 2, y = 3.5, label = leg, hjust = 0) +
  ggtitle("Average error of four different timeseries forecasting methods
M3 Forecasting Competition data") +
  labs(x = "Model, or ensemble of models
(further to the left means better overall performance)",
  y = "Mean scaled absolute error\\n(smaller numbers are better)") +
  theme_grey(9)
  }
}
\author{
Peter Ellis
}
\keyword{datasets}
