% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{plot.xgbar}
\alias{plot.xgbar}
\title{Plot xgbar object}
\usage{
\method{plot}{xgbar}(x, ...)
}
\arguments{
\item{x}{An object created by \code{xgbar}}

\item{...}{Additional arguments passed through to \code{plot()}}
}
\description{
plot method for an object created by xgbar
}
\examples{
model <- xgbar(AirPassengers)
plot(model)
}
\seealso{
\code{\link{xgbar}}
}
\author{
Peter Ellis
}
