% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{plot.xgbts}
\alias{plot.xgbts}
\title{Plot xgbts object}
\usage{
\method{plot}{xgbts}(x, ...)
}
\arguments{
\item{x}{An object created by \code{xgbts}}

\item{...}{Additional arguments passed through to \code{plot()}}
}
\description{
plot method for an object created by xgbts
}
\examples{
model <- xgbts(AirPassengers)
plot(model)
}
\author{
Peter Ellis
}

