% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-doc.R
\docType{data}
\name{seaice_ts}
\alias{seaice_ts}
\title{Arctic sea ice}
\format{An object of class \code{ts} of length 10647.}
\source{
National Snow and Ice Data Center, \url{https://nsidc.org/data/docs/noaa/g02135_seaice_index/#daily_data_files"}
}
\usage{
seaice_ts
}
\description{
Extent in millions of square kilometres of sea ice in the Arctic from 21 August 1987 to 24 November 2016.
}
\examples{
plot(seaice_ts)
}
\keyword{datasets}
