% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{summary.xgbar}
\alias{print.summary.xgbar}
\alias{summary.xgbar}
\title{Summary of an xgbar object}
\usage{
\method{summary}{xgbar}(object, ...)
}
\arguments{
\item{object}{An object created by \code{\link{xgbar}}}

\item{...}{Ignored.}
}
\description{
summary method for an object created by xgbar
}
\examples{
\dontrun{
# Half-hourly electricity demand in England and Wales, takes a few minutes
electricity_model <- xgbar(taylor)
summary(electricity_model)
electricity_fc <- forecast(electricity_model, h = 500)
plot(electricity_fc)
}
}
\author{
Peter Ellis
}
\seealso{
\code{\link{xgbar}}
}

