% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\docType{data}
\name{Tcomp_results}
\alias{Tcomp_results}
\title{Tourism forecasting results}
\format{A data frame with 60 rows and three columns.}
\usage{
Tcomp_results
}
\description{
Summary data from four models, and 11 combinations of models, against the data from the 2010 tourism forecasting competition.
}
\details{
Full details of how this was generated are in the Vignette.  This shows the average mean absolute scaled error 
(MASE) from using \code{xgbar} (x), \code{auto.arima} (a), \code{nnetar} (n) and \code{thetaf} (f) to generate forecasts of 1,311 tourism data series.

\itemize{
\item MASE A mean mean absolute squared error
\item model model, or ensemble of models, to which the MASE applies
\item Frequency The frequency of the subset of data from which the mean MASE was calculated.
}
}
\examples{
if(require(ggplot2)){
leg <- "f: Theta; forecast::thetaf\\na: ARIMA; forecast::auto.arima
n: Neural network; forecast::nnetar\\nx: Extreme gradient boosting; forecastxgb::xgbar"

ggplot(Tcomp_results, aes(x = model, y =  MASE, colour = Frequency, label = model)) +
  geom_text(size = 4) +
  geom_line(aes(x = as.numeric(model)), alpha = 0.25) +
  annotate("text", x = 2, y = 3.5, label = leg, hjust = 0) +
  ggtitle("Average error of four different timeseries forecasting methods
2010 Tourism Forecasting Competition data") +
  labs(x = "Model, or ensemble of models
(further to the left means better overall performance)",
  y = "Mean scaled absolute error\\n(smaller numbers are better)") +
  theme_grey(9)
  }
}
\author{
Peter Ellis
}
\keyword{datasets}

