% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{xgbar_importance}
\alias{xgbar_importance}
\title{Show importance of features in a xgbar model}
\usage{
xgbar_importance(object, ...)
}
\arguments{
\item{object}{An object of class \code{xgbar}, usually created with \code{xgbar()}}

\item{...}{Extra parameters passed through to \code{xgb.importance}}
}
\value{
A \code{data.table} of the features used in the model with their average gain 
(and their weight for boosted tree model) in the model.
}
\description{
This is a light wrapper for \code{xgboost::xbg.importance} to make it easier to use with objects of class \code{xgbar}
}
\seealso{
\code{\link[xgboost]{xgb.importance}}, \code{\link{summary.xgbar}}, \code{\link{xgbar}}.
}
\author{
Peter Ellis
}
