% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.xgbar.R
\name{forecast.xgbar}
\alias{forecast.xgbar}
\title{Forecasting using xgboost models}
\usage{
\method{forecast}{xgbar}(object, h = NULL, xreg = NULL, ...)
}
\arguments{
\item{object}{An object of class "\code{xgbar}".  Usually the result of a call to \code{\link{xgbar}}.}

\item{h}{Number of periods for forecasting.  If \code{xreg} is provided, the number of rows of \code{xreg} will be 
used and \code{h} is ignored with a warning.  If both \code{h} and \code{xreg} are \code{NULL} then 
\code{h = ifelse(frequency(object$y) > 1, 2 * frequency(object$y), 10)}}

\item{xreg}{Future values of regression variables.}

\item{...}{Ignored.}
}
\value{
An object of class \code{forecast}
}
\description{
Returns forecasts and other information for xgboost timeseries modesl fit with \code{xbgts}
}
\examples{
# Australian monthly gas production
gas_model <- xgbar(gas)
summary(gas_model)
gas_fc <- forecast(gas_model, h = 12)
plot(gas_fc)
}
\author{
Peter Ellis
}
\seealso{
\code{\link{xgbar}}, \code{\link[forecast]{forecast}}
}

