% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbts.R
\name{forecast.xgbts}
\alias{forecast.xgbts}
\title{Forecasting using xgboost models}
\usage{
\method{forecast}{xgbts}(object, h = ifelse(frequency(object$y) > 1, 2 *
  frequency(object$y), 10), xreg = NULL)
}
\description{
Returns forecasts and other information for xgboost timeseries modesl fit with \code{xbgts}
}

