% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbts.R
\name{forecast.xgbts}
\alias{forecast.xgbts}
\title{Forecasting using xgboost models}
\usage{
\method{forecast}{xgbts}(object, h = NULL, xreg = NULL, ...)
}
\arguments{
\item{object}{An object of class "\code{xgbts}".  Usuall the result of a call to \code{\link{xgbts}}.}

\item{h}{Number of periods for forecasting}

\item{xreg}{Future values of regression variables.}

\item{...}{Ignored.}
}
\value{
An object of class \code{forecast}
}
\description{
Returns forecasts and other information for xgboost timeseries modesl fit with \code{xbgts}
}
\examples{
# Australian monthly gas production
gas_model <- xgbts(gas)
summary(gas_model)
gas_fc <- forecast(gas_model, h = 12)
plot(gas_fc)
}
\author{
Peter Ellis
}
\seealso{
\code{\link{xgbts}}, \code{\link[forecast]{forecast}}
}

