% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.xgbts}
\alias{plot.xgbts}
\title{Plot xgbts object}
\usage{
plot.xgbts(object, ...)
}
\arguments{
\item{object}{An object created by \code{xgbts}}

\item{...}{Additional arguments passed through to \code{plot()}}
}
\description{
plot method for an object created by xgbts
}

