% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{summary.xgbts}
\alias{print.summary.xgbts}
\alias{summary.xgbts}
\title{Summary of an xgbts object}
\usage{
\method{summary}{xgbts}(object, ...)
}
\arguments{
\item{object}{An object created by \code{xgbts}}

\item{...}{Ignored.}
}
\description{
summary method for an object created by xgbts
}

