% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbts.R
\name{xgbts}
\alias{xgbts}
\title{xgboost time series modelling}
\usage{
xgbts(y, xreg = NULL, maxlag = 2 * frequency(y), nrounds = 100,
  verbose = FALSE, ...)
}
\arguments{
\item{nrounds}{Maximum number of iterations in cross validation to determine}

\item{nfold}{Number of equal size subsamples during cross validation}
}
\value{
An object of class \code{xgbts}
}
\description{
Fit a model to a time series using xgboost
}
\author{
Peter Ellis
}

